refactor(controller): extract wav entries loader
